Delete public/Day35 directory
Deleted Day35 folder to fix case-sensitivity conflict. There are two file in folder by name Day35 and day35 due to which it was causing the conflict. And after check carefully the Day35 was not useful and day35 is the needed file. So, the Day35 is been deleted to avoid conflicts further.